+++ packages/rocketh-web/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @rocketh/web
 
+## 0.17.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @rocketh/core@0.17.2
+  - rocketh@0.17.3
+
 ## 0.17.2
 
 ### Patch Changes